fix: repeat algo on completion. (#83)
* fix: repeat algo on completion.

* fix: rename file
