+++ src/components/buttons/hooks/index.ts
@@ -1,0 +1,2 @@
+import useOnClick from './useOnClick';
+export { useOnClick };
